Corrijo type ID a Long en Entities: Docente, Preceptor, Tutor

+++ src/main/java/com/grup14/luterano/entities/Preceptor.java
@@ -10,5 +10,5 @@ import jakarta.persistence.Id;
 public class Preceptor extends Persona {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 }
